+++ cpp_src/cmd/reindexer_tool/repair_tool.cc
@@ -73,7 +73,7 @@ Error RepairTool::repairNamespace(IDataStorage* storage, const std::string& stor
 		ns.LoadFromStorage(dummyCtx);
 	} catch (const Error& err) {
 		std::cout << "Namespace was not repaired: " << err.what() << ". Should it be deleted? y/N" << std::endl;
-		while (true) {
+		for (;;) {
 			std::string input;
 			std::getline(std::cin, input);
 			std::transform(input.begin(), input.end(), input.begin(), [](char c) { return std::tolower(c); });