+++ cpp_src/core/querystat.cc
@@ -7,23 +7,21 @@
 namespace reindexer {
 
 template <void (PerfStatCounterST::*hitFunc)(std::chrono::microseconds)>
-void QueriesStatTracer::hit(const Query &q, std::chrono::microseconds time) {
-	WrSerializer serNorm, serNonNorm;
-	const string_view sqlq{q.GetSQL(serNorm, true).Slice()};
+void QueriesStatTracer::hit(const QuerySQL &sql, std::chrono::microseconds time) {
 	std::unique_lock<std::mutex> lck(mtx_);
-	const auto it = stat_.find(sqlq);
+	const auto it = stat_.find(sql.normolized);
 	if (it == stat_.end()) {
-		(stat_.emplace(string(sqlq), Stat(q.GetSQL(serNonNorm, false).Slice())).first->second.*hitFunc)(time);
+		(stat_.emplace(string(sql.normolized), Stat(sql.nonNormolized)).first->second.*hitFunc)(time);
 	} else {
 		const auto maxTime = it->second.MaxTime();
 		(it->second.*hitFunc)(time);
 		if (it->second.MaxTime() > maxTime) {
-			it->second.longestQuery = string(q.GetSQL(serNonNorm, false).Slice());
+			it->second.longestQuery = string(sql.nonNormolized);
 		}
 	}
 };
-template void QueriesStatTracer::hit<&PerfStatCounterST::Hit>(const Query &, std::chrono::microseconds);
-template void QueriesStatTracer::hit<&PerfStatCounterST::LockHit>(const Query &, std::chrono::microseconds);
+template void QueriesStatTracer::hit<&PerfStatCounterST::Hit>(const QuerySQL &, std::chrono::microseconds);
+template void QueriesStatTracer::hit<&PerfStatCounterST::LockHit>(const QuerySQL &, std::chrono::microseconds);
 
 const std::vector<QueryPerfStat> QueriesStatTracer::Data() {
 	std::unique_lock<std::mutex> lck(mtx_);